+++ core/client/fs/src/main/java/alluxio/client/block/stream/BlockWorkerClient.java
@@ -11,73 +11,31 @@
 
 package alluxio.client.block.stream;
 
-import alluxio.Configuration;
-import alluxio.PropertyKey;
-import alluxio.grpc.BlockWorkerGrpc;
 import alluxio.grpc.ReadRequest;
 import alluxio.grpc.ReadResponse;
 import alluxio.grpc.WriteRequest;
 import alluxio.grpc.WriteResponse;
-import alluxio.util.grpc.GrpcChannel;
-import alluxio.util.grpc.GrpcChannelBuilder;
-import alluxio.util.network.NettyUtils;
 
-import io.grpc.ConnectivityState;
 import io.grpc.stub.StreamObserver;
-import io.netty.channel.EventLoopGroup;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import java.io.Closeable;
-import java.io.IOException;
-import java.net.InetSocketAddress;
 import java.net.SocketAddress;
 import java.util.Iterator;
-import java.util.concurrent.TimeUnit;
 
 import javax.security.auth.Subject;
 
 /**
  * gRPC client for worker communication.
  */
-public class BlockWorkerClient implements Closeable {
-  private static final Logger LOGGER = LoggerFactory.getLogger(BlockWorkerClient.class.getName());
-
-  private static final long KEEPALIVE_TIMEOUT_MS =
-      Configuration.getMs(PropertyKey.USER_NETWORK_NETTY_TIMEOUT_MS);
-  private static final long GRPC_FLOWCONTROL_WINDOW =
-      Configuration.getMs(PropertyKey.USER_NETWORK_GRPC_FLOWCONTROL_WINDOW);
-  private static final long MAX_INBOUND_MESSAGE_SIZE =
-      Configuration.getMs(PropertyKey.USER_NETWORK_GRPC_MAX_INBOUND_MESSAGE_SIZE);
-  private static final EventLoopGroup WORKER_GROUP = NettyUtils
-      .createEventLoop(NettyUtils.USER_CHANNEL_TYPE,
-          Configuration.getInt(PropertyKey.USER_NETWORK_NETTY_WORKER_THREADS),
-          "netty-client-worker-%d", true);
-
-  private GrpcChannel mChannel;
-  private BlockWorkerGrpc.BlockWorkerBlockingStub mBlockingStub;
-  private BlockWorkerGrpc.BlockWorkerStub mAsyncStub;
-
+public interface BlockWorkerClient extends Closeable {
   /**
    * Builder for the block worker client.
    */
-  public static class Builder {
-    private final GrpcChannelBuilder mChannelBuilder;
-
-    /**
-     * Creates a {@link Builder} for {@link BlockWorkerClient}.
-     * @param channelBuilder a gRPC channel builder
-     */
-    public Builder(GrpcChannelBuilder channelBuilder) {
-      mChannelBuilder = channelBuilder;
-    }
-
+  interface Builder {
     /**
      * @return a new {@link BlockWorkerClient}
      */
-    public BlockWorkerClient build() {
-      return new BlockWorkerClient(mChannelBuilder.build());
-    }
+    BlockWorkerClient build();
   }
 
   /**
@@ -87,51 +45,14 @@ public class BlockWorkerClient implements Closeable {
    * @param address the address of the worker
    * @return the builder for the client
    */
-  public static Builder getBuilder(Subject subject, SocketAddress address) {
-    return new Builder(GrpcChannelBuilder
-        .forAddress(address)
-        .channelType(NettyUtils.getClientChannelClass(
-            !(address instanceof InetSocketAddress)))
-        .group(WORKER_GROUP)
-        .usePlaintext(true)
-        .keepAliveTimeout(KEEPALIVE_TIMEOUT_MS, TimeUnit.MILLISECONDS)
-        .maxInboundMessageSize((int) MAX_INBOUND_MESSAGE_SIZE)
-        .flowControlWindow((int) GRPC_FLOWCONTROL_WINDOW));
-  }
-
-  /**
-   * Creates a client instance for communicating with block worker.
-   *
-   * @param channel the gRPC channel
-   */
-  public BlockWorkerClient(GrpcChannel channel) {
-    mChannel = channel;
-    mBlockingStub = BlockWorkerGrpc.newBlockingStub(mChannel);
-    mAsyncStub = BlockWorkerGrpc.newStub(mChannel);
+  static Builder getBuilder(Subject subject, SocketAddress address) {
+    return DefaultBlockWorkerClient.getBuilder(subject, address);
   }
 
   /**
    * @return whether the client is shutdown
    */
-  public boolean isShutdown() {
-    ConnectivityState state = mChannel.getState(false);
-    return state == ConnectivityState.SHUTDOWN;
-  }
-
-  @Override
-  public void close() throws IOException {
-    mChannel.shutdown();
-    try {
-      if (!mChannel.awaitTermination(
-          Configuration.getMs(PropertyKey.WORKER_NETWORK_NETTY_SHUTDOWN_TIMEOUT),
-          TimeUnit.MILLISECONDS)) {
-        LOGGER.warn("Failed to close gRPC channel for block worker.");
-      }
-    } catch (InterruptedException e) {
-      Thread.currentThread().interrupt();
-      throw new RuntimeException(e);
-    }
-  }
+  boolean isShutdown();
 
   /**
    * Writes a block to the worker.
@@ -139,9 +60,7 @@ public class BlockWorkerClient implements Closeable {
    * @param responseObserver the stream observer for the server response
    * @return the stream observer for the client request
    */
-  public StreamObserver<WriteRequest> writeBlock(StreamObserver<WriteResponse> responseObserver) {
-    return mAsyncStub.writeBlock(responseObserver);
-  }
+  StreamObserver<WriteRequest> writeBlock(StreamObserver<WriteResponse> responseObserver);
 
   /**
    * Reads a block from the worker.
@@ -149,7 +68,5 @@ public class BlockWorkerClient implements Closeable {
    * @param request the read request
    * @return the streamed response from server
    */
-  public Iterator<ReadResponse> readBlock(final ReadRequest request) {
-    return mBlockingStub.readBlock(request);
-  }
+  Iterator<ReadResponse> readBlock(final ReadRequest request);
 }
